refactor: 타입 선언 정리
