+++ src/apis/getChatRoomList.ts
@@ -1,15 +1,6 @@
 import chatRoomList from '@/assets/data/chatroomList.json';
-import { MessageDto, UserDto } from '@/pages/ChatRoom/dto';
-
-export interface ChatRoomDto {
-	chatRoomId: number;
-	chatRoomName: string | null;
-	joinedUsers: UserDto[];
-	latestMessage: MessageDto;
-}
-
-type ChatRoomResponse = ChatRoomDto[];
+import { ChatRoomDto } from '@/apis/dto';
 
 export const getChatRoomList = () => {
-	return chatRoomList as ChatRoomResponse;
+	return chatRoomList as ChatRoomDto[];
 };
